tidy up crossbar router and fixture
